requirements.txt
boto3
attrs
python-packer
moto
pytest
